moving router helper code to it's own module.
And fixing issue with modal
